+++ app/assets/scripts/selfie/campaign.js
@@ -938,14 +938,15 @@ function( angular , c6State  , PaginatedListState                    ,
             };
         }])
 
-        .controller('SelfieCampaignSponsorController', ['$scope','CollateralService',
-        function                                       ( $scope , CollateralService ) {
+        .controller('SelfieCampaignSponsorController', ['$scope','CollateralService','c6State',
+        function                                       ( $scope , CollateralService , c6State ) {
             var AppCtrl = $scope.AppCtrl,
                 SelfieCampaignSponsorCtrl = this,
                 SelfieCampaignCtrl = $scope.SelfieCampaignCtrl,
                 advertiser = SelfieCampaignCtrl.advertiser,
                 defaultLogo = advertiser.defaultLogos && advertiser.defaultLogos.square,
-                card = SelfieCampaignCtrl.card;
+                card = SelfieCampaignCtrl.card,
+                websiteLogo = card.collateral.logoType === 'website' && card.collateral.logo;
 
             function typeify(name) {
                 switch (name) {
@@ -981,11 +982,11 @@ function( angular , c6State  , PaginatedListState                    ,
                         label: option
                     });
                     return result;
-                }, defaultLogo ?
+                }, websiteLogo ?
                 [{
-                    type: 'account',
-                    label: 'Account Default',
-                    src: defaultLogo
+                    type: 'website',
+                    label: 'Website Default',
+                    src: websiteLogo
                 }] : [])
                 .concat(SelfieCampaignCtrl.logos
                     .map(function(logo) {
@@ -1011,26 +1012,139 @@ function( angular , c6State  , PaginatedListState                    ,
             this.previouslyUploadedLogo = undefined;
 
             this.links = [
-                    'Website','Sharing','Facebook','Twitter',
+                    'Sharing','Facebook','Twitter',
                     'Instagram','YouTube','Pinterest'
                 ].map(function(name) {
-                    var href = card.links[name] || null,
-                        cssClass = name.toLowerCase(),
-                        isWebsite = cssClass === 'website';
+                    var href = card.links[name] || '',
+                        cssClass = name.toLowerCase();
 
                     return {
-                        cssClass: isWebsite ? 'link' : cssClass,
+                        cssClass: cssClass,
                         name: name,
                         href: href,
-                        required: isWebsite
+                        required: false
                     };
                 });
 
+            this.website = card.links.Website;
             this.sharing = card.shareLinks.facebook;
+            this.bindShareToWebsite = !this.sharing;
+            this.loadingSiteData = false;
+            this.siteDataFailure = false;
+            this.siteDataSuccess = false;
+            this.hasImported = !!this.website || !!this.links.filter(function(link) {
+                return !!link.href;
+            }).length;
 
+            // this is called on-focus of website input,
+            // the allowImport flag will show the import
+            // button when appropriate
+            this.checkImportability = function() {
+                this.allowImport = this.hasImported;
+            };
+
+            // return a url with a protocol or undefined
+            this.validateWebsite = function () {
+                var website = this.website;
+
+                if (!website) { return; }
+
+                if (!(/^http:\/\/|https:\/\//).test(website)) {
+                    website = 'http://' + website;
+                }
+
+                return website;
+            };
+
+            // this is called when website data is successfully fetched,
+            // it updates props on the Ctrl not the actual card
+            this.setWebsiteData = function(data) {
+                var links = data.links,
+                    logo = data.images && data.images.profile;
+
+                if (links) {
+                    this.links.forEach(function(link) {
+                        var name = link.name.toLowerCase();
+                        link.href = (!!links[name] || links[name] === null) ?
+                            links[name] : link.href;
+                    });
+                }
+
+                if (logo) {
+                    if (this.logoOptions[0].type === 'website') {
+                        this.logoOptions[0].src = logo;
+                    } else {
+                        this.logoOptions = [{
+                            type: 'website',
+                            label: 'Website Default',
+                            src: logo
+                        }].concat(this.logoOptions);
+                    }
+
+                    this.logoType = this.logoOptions[0];
+                }
+            };
+
+            // there's a UI button for this, it's only available once
+            // a website is set and the user clicks into the input
+            this.importWebsite = function() {
+                var website = this.validateWebsite();
+                if (!website) { return; }
+
+                this.allowImport = false;
+
+                c6State.goTo('Selfie:Campaign:Website', [{website: website}]);
+            };
+
+            // this is the method that is called from modal
+            this.saveWebsiteData = function(data) {
+                this.setWebsiteData(data);
+                this.updateLinks();
+            };
+
+            // this is called on-blur of website input,
+            // we only automatically call CollateralService
+            // the first time a user enters a website,
+            // after that they need to use the "import" button
+            this.checkWebsite = function() {
+                var website = this.validateWebsite(),
+                    currentWebsite = card.links.Website;
+
+                this.siteDataSuccess = false;
+                this.siteDataFailure = false;
+
+                if (!website || this.allowImport || website === currentWebsite) {
+                    this.updateLinks();
+                    return;
+                }
+
+                this.loadingSiteData = true;
+
+                CollateralService.websiteData(website)
+                    .then(function(data) {
+                        SelfieCampaignSponsorCtrl.siteDataSuccess = {
+                            logo: !!data.images.profile,
+                            links: Object.keys(data.links).filter(function(key) {
+                                return !!data.links[key];
+                            }).length
+                        };
+                        SelfieCampaignSponsorCtrl.setWebsiteData(data);
+                    })
+                    .catch(function() {
+                        SelfieCampaignSponsorCtrl.siteDataFailure = true;
+                    })
+                    .finally(function() {
+                        SelfieCampaignSponsorCtrl.updateLinks();
+                        SelfieCampaignSponsorCtrl.hasImported = true;
+                        SelfieCampaignSponsorCtrl.loadingSiteData = false;
+                    });
+            };
+
+            // this is called on-blur of all links inputs (except website)
+            // it's also called during importing/updating of website
             this.updateLinks = function() {
-                var sharing = SelfieCampaignSponsorCtrl.sharing,
-                    shareLink = AppCtrl.validUrl.test(sharing) ? sharing : 'http://' + sharing;
+                var website = SelfieCampaignSponsorCtrl.validateWebsite(),
+                    sharing, shareLink;
 
                 SelfieCampaignSponsorCtrl.links.forEach(function(link) {
                     if (link.href && link.href === card.links[link.name]) { return; }
@@ -1042,6 +1156,16 @@ function( angular , c6State  , PaginatedListState                    ,
                         card.links[link.name] = undefined;
                     }
                 });
+
+                card.links.Website = website;
+
+                // ensure that sharing link is updated if appropriate
+                if (SelfieCampaignSponsorCtrl.bindShareToWebsite) {
+                    SelfieCampaignSponsorCtrl.sharing = website;
+                }
+
+                sharing = SelfieCampaignSponsorCtrl.sharing;
+                shareLink = AppCtrl.validUrl.test(sharing) ? sharing : 'http://' + sharing;
 
                 if (sharing === card.shareLinks.facebook) { return; }
 
@@ -1070,7 +1194,8 @@ function( angular , c6State  , PaginatedListState                    ,
                 var selectedType = SelfieCampaignSponsorCtrl.logoType.type;
 
                 card.collateral.logo = newLogo;
-                card.collateral.logoType = /file|url/.test(selectedType) ? selectedType : undefined;
+                card.collateral.logoType = /file|url|website/.test(selectedType) ?
+                    selectedType : undefined;
             });
 
             // we do very different things depending on what the
@@ -1090,6 +1215,7 @@ function( angular , c6State  , PaginatedListState                    ,
                     }
                     break;
                 case 'custom':
+                case 'website':
                     Ctrl.logo = Ctrl.logoType.src;
                     break;
                 case 'account':
@@ -1113,6 +1239,76 @@ function( angular , c6State  , PaginatedListState                    ,
             });
 
             $scope.$on('SelfieCampaignWillSave', this.updateLinks);
+        }])
+
+        .config(['c6StateProvider',
+        function( c6StateProvider ) {
+            c6StateProvider.state('Selfie:Campaign:Website', [function() {
+                this.templateUrl = 'views/selfie/campaigns/edit/website_scraping_dialog.html';
+                this.controller = 'SelfieCampaignWebsiteController';
+                this.controllerAs = 'SelfieCampaignWebsiteCtrl';
+            }]);
+        }])
+
+        .controller('SelfieCampaignWebsiteController', ['c6State','cState','CollateralService',
+        function                                       ( c6State , cState , CollateralService ) {
+            var SelfieCampaignWebsiteCtrl = this;
+
+            Object.defineProperties(this, {
+                data: {
+                    get: function() {
+                        var logo = this.logo || {},
+                            links = this.links || [];
+
+                        return {
+                            links: links.reduce(function(result, link) {
+                                if (link.selected) { result[link.name] = link.href; }
+                                return result;
+                            }, {}),
+                            images: {
+                                profile: logo.selected && logo.href
+                            }
+                        };
+                    }
+                }
+            });
+
+            this.close = function() {
+                c6State.goTo(cState.cParent.cName);
+            };
+
+            this.initWithModel = function(model) {
+                this.loading = true;
+
+                CollateralService.websiteData(model.website).then(function(data) {
+                    SelfieCampaignWebsiteCtrl.logo = {
+                        href: data.images.profile,
+                        selected: true
+                    };
+
+                    SelfieCampaignWebsiteCtrl.links = (function() {
+                        var options = [];
+
+                        forEach(data.links, function(link, key) {
+                            options.push({
+                                cssClass: ((/instagram|google/).test(key) ?
+                                    key : key + '-square'),
+                                name: key,
+                                href: link,
+                                selected: true
+                            });
+                        });
+
+                        return options;
+                    }());
+
+                    SelfieCampaignWebsiteCtrl.loading = false;
+
+                }, function() {
+                    SelfieCampaignWebsiteCtrl.loading = false;
+                    SelfieCampaignWebsiteCtrl.error = true;
+                });
+            };
         }])
 
         .controller('SelfieCampaignVideoController', ['$injector','$scope','SelfieVideoService',